Create MNISTclassification.py
The task was to create a neural network capable of classifying handwritten digits from the MNIST dataset. The dataset is made up of 70,000 images, of which 60,000 images are 60 000 images are for learning and 10 000 images are for testing. Each image represents one digit from 0 to 9, which is represented using shades of grey and in a size of 28x28.